Extracted skin logic to a separate definition
