+++ planet/Objects.ocd/Clonk.ocd/Skins.ocd/DefCore.txt
@@ -1,0 +1,5 @@
+[DefCore]
+id=Clonk_Skins
+Version=8,0
+Category=C4D_StaticBack
+HideInCreator=true
