+++ planet/Objects.ocd/Clonk.ocd/Skins.ocd/StringTblDE.txt
@@ -1,0 +1,2 @@
+Skin=Skin
+SkinHelp=Aussehen des Clonks.
