+++ planet/Objects.ocd/Clonk.ocd/Skins.ocd/StringTblUS.txt
@@ -1,0 +1,2 @@
+Skin=Skin
+SkinHelp=How the clonk looks.
